requirements.txt
appnope==0.1.0
astroid==2.3.3
backcall==0.1.0
cycler==0.10.0
decorator==4.4.2
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
jupyter-client==6.1.2
jupyter-core==4.6.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
numpy-quaternion==2019.12.11.22.25.52
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
pyzmq==19.0.0
six==1.14.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.9
wrapt==1.11.2
